requirements.txt
numpy
scipy
python-dotenv
PyYAML
Jinja2

pymumble
resampy

# libsamplerate wrapper
samplerate

cryptography
aiofiles

# rabbitmq
aio-pika

# continuous wavelet transform
pywavelets

matplotlib